pyproject.toml
[project]
name = "renewable-lcoe-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.5",
    "gradio>=5",
    "ipykernel>=6.29.5",
    "pandas>=2.2.3",
    "pip>=24.3.1",
    "plotly-express>=0.4.1",
    "polars>=1.13.1",
    "pyarrow>=18.0.0",
    "pydantic>2.9.0",
    "pyxirr>=0.10.6",
    "ruff>=0.7.4",
    "scipy>=1.14.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
